VM properties dialog: convert disks fields into components
The existing code was not using React components enough. Converting
before heavy changes.
