+++ web/html/src/manager/virtualization/guests/create/guests-create.js
@@ -10,7 +10,7 @@ const { getOrderedItemsFromModel } = require('components/input/FormMultiInput');
 const { GuestProperties } = require('../guest-properties');
 const { VirtualizationGuestActionApi } = require('../virtualization-guest-action-api');
 const GuestNicsPanel = require('../properties/guest-nics-panel');
-const GuestDisksPanel = require('../properties/guest-disks-panel');
+const DiskUtils = require('../properties/disk-utils');
 const Functions = require('utils/functions');
 
 type Props = {
@@ -33,7 +33,7 @@ class GuestsCreate extends React.Component<Props, State> {
 
     // Diff the model with the initial one to avoid changing disks if user hasn't touched them.
     const disks = getOrderedItemsFromModel(model, 'disk')
-      .map(index => GuestDisksPanel.getRequestParams(model, index));
+      .map(index => DiskUtils.getRequestParams(model, index));
 
     const filteredProps = ['disk', 'network', 'vmType'];
     return Object.assign(
