+++ web/html/src/manager/virtualization/guests/edit/guests-edit.js
@@ -12,7 +12,7 @@ const { Loading } = require('components/utils/Loading');
 const { getOrderedItemsFromModel } = require('components/input/FormMultiInput');
 const { GuestProperties } = require('../guest-properties');
 const GuestNicsPanel = require('../properties/guest-nics-panel');
-const GuestDisksPanel = require('../properties/guest-disks-panel');
+const DiskUtils = require('../properties/disk-utils');
 const { VirtualizationGuestActionApi } = require('../virtualization-guest-action-api');
 const { VirtualizationGuestDefinitionApi } = require('../virtualization-guest-definition-api');
 const Functions = require('utils/functions');
@@ -37,7 +37,7 @@ class GuestsEdit extends React.Component<Props> {
       vmType: definition.type,
     },
     GuestNicsPanel.getModelFromDefinition(definition),
-    GuestDisksPanel.getModelFromDefinition(definition));
+    DiskUtils.getModelFromDefinition(definition));
   }
 
   static getRequestParameterFromModel(model: Object, initialModel: Object) {
@@ -56,7 +56,7 @@ class GuestsEdit extends React.Component<Props> {
     const newDiskProps = Object.entries(model).filter(entry => entry[0].startsWith('disk'));
     const disks = !_isEqual(initialDiskProps, newDiskProps)
       ? getOrderedItemsFromModel(model, 'disk')
-        .map(index => GuestDisksPanel.getRequestParams(model, index))
+        .map(index => DiskUtils.getRequestParams(model, index))
       : [];
 
     const disksParams = disks.length !== 0 ? { disks } : undefined;
